adding window running trait

+++ crystal/src/classes/conf.rs
@@ -1,6 +1,6 @@
 pub struct GameConfig {
     pub game_name: String,
-    pub engine_version: [u8; 3],
+    pub engine_version: [u16; 3],
     pub window_title: String,
     pub window_width: u32,
     pub window_height: u32,
